refactor(scripts): move App Store Connect tools to scripts/asc/
- Package asc_*.py and asc_client under scripts/asc/ with imports scripts.asc.*
- Fix asc_client sys.path and asc_strict_screenshot_sync repo_root depth
- Update workflows, docs, tests (including mock.patch targets), check_ios_version_lineage

Made-with: Cursor
