+++ .github/workflows/ios-submit-review.yml
@@ -122,7 +122,7 @@ jobs:
           APPSTORE_PRIVATE_KEY: /tmp/appstore_key.p8
           PREFERRED_VERSION: ${{ steps.version.outputs.ios_version }}
         run: |
-          python scripts/asc_resolve_version.py \
+          python scripts/asc/asc_resolve_version.py \
             --preferred-version "$PREFERRED_VERSION" \
             --create-if-needed \
             --auto-next-patch \
@@ -161,7 +161,7 @@ jobs:
           APPSTORE_PRIVATE_KEY: /tmp/appstore_key.p8
           IOS_VERSION: ${{ steps.asc_version.outputs.selected_version }}
         run: |
-          STATE_JSON="$(python scripts/asc_poll_version_state.py --version "$IOS_VERSION" --json 2>/dev/null)" || true
+          STATE_JSON="$(python scripts/asc/asc_poll_version_state.py --version "$IOS_VERSION" --json 2>/dev/null)" || true
           STATE="$(echo "$STATE_JSON" | python -c 'import json,sys; print(json.loads(sys.stdin.read()).get("state","UNKNOWN"))' 2>/dev/null)" || STATE="UNKNOWN"
           echo "state=$STATE" >> "$GITHUB_OUTPUT"
           echo "ASC version state: $STATE"
@@ -204,8 +204,8 @@ jobs:
           fi
           python - <<'PY'
           import os
-          from scripts.asc_client import ASCClient
-          from scripts.asc_submit_for_review import attach_build, select_valid_build_id
+          from scripts.asc.asc_client import ASCClient
+          from scripts.asc.asc_submit_for_review import attach_build, select_valid_build_id
 
           client = ASCClient.from_env(timeout=30)
           app_id = os.environ["APP_ID"]
@@ -237,7 +237,7 @@ jobs:
           if [[ "${{ inputs.attach_subscriptions }}" == "true" ]]; then
             EXTRA_FLAGS="--skip-build-check"
           fi
-          python scripts/asc_verify_ready.py \
+          python scripts/asc/asc_verify_ready.py \
             --version "$IOS_VERSION" \
             --locale "$LOCALE" \
             $EXTRA_FLAGS \
@@ -273,7 +273,7 @@ jobs:
             exit 1
           fi
           # Read current ASC state first so reruns are safe/idempotent.
-          STATE_JSON="$(python scripts/asc_poll_version_state.py --version "$IOS_VERSION" --json)"
+          STATE_JSON="$(python scripts/asc/asc_poll_version_state.py --version "$IOS_VERSION" --json)"
           echo "ASC state (pre-submit): $STATE_JSON"
           STATE="$(python -c 'import json,sys; print(json.loads(sys.argv[1])["state"])' "$STATE_JSON")"
 
@@ -292,7 +292,7 @@ jobs:
           if [[ "$ATTACH_SUBSCRIPTIONS" == "true" ]]; then
             SUBMIT_ARGS+=(--attach-subscriptions)
           fi
-          python scripts/asc_submit_for_review.py "${SUBMIT_ARGS[@]}"
+          python scripts/asc/asc_submit_for_review.py "${SUBMIT_ARGS[@]}"
 
       - name: Upload version resolution report
         if: always()
