+++ .github/workflows/native-release.yml
@@ -626,7 +626,7 @@ jobs:
           APPSTORE_PRIVATE_KEY: ${{ secrets.APPSTORE_PRIVATE_KEY }}
           PREFERRED_VERSION: ${{ steps.versions.outputs.ios_version }}
         run: |
-          python scripts/asc_resolve_version.py \
+          python scripts/asc/asc_resolve_version.py \
             --preferred-version "$PREFERRED_VERSION" \
             --create-if-needed \
             --json-out /tmp/asc_version.json
@@ -653,7 +653,7 @@ jobs:
           APPSTORE_ISSUER_ID: ${{ secrets.APPSTORE_ISSUER_ID }}
           APPSTORE_PRIVATE_KEY: /tmp/appstore_key.p8
         run: |
-          STATE_JSON="$(python scripts/asc_poll_version_state.py --version "${{ steps.asc_version.outputs.selected_version }}" --json 2>/dev/null)" || true
+          STATE_JSON="$(python scripts/asc/asc_poll_version_state.py --version "${{ steps.asc_version.outputs.selected_version }}" --json 2>/dev/null)" || true
           STATE="$(echo "$STATE_JSON" | python -c 'import json,sys; print(json.loads(sys.stdin.read()).get("state","UNKNOWN"))' 2>/dev/null)" || STATE="UNKNOWN"
           echo "state=$STATE" >> "$GITHUB_OUTPUT"
           echo "ASC version state before metadata upload: $STATE"
@@ -689,7 +689,7 @@ jobs:
           APPSTORE_PRIVATE_KEY: /tmp/appstore_key.p8
           IOS_VERSION: ${{ steps.asc_version.outputs.selected_version }}
         run: |
-          STATE_JSON="$(python scripts/asc_poll_version_state.py --version "$IOS_VERSION" --json 2>/dev/null)" || true
+          STATE_JSON="$(python scripts/asc/asc_poll_version_state.py --version "$IOS_VERSION" --json 2>/dev/null)" || true
           STATE="$(echo "$STATE_JSON" | python -c 'import json,sys; print(json.loads(sys.stdin.read()).get("state","UNKNOWN"))' 2>/dev/null)" || STATE="UNKNOWN"
           echo "state=$STATE" >> "$GITHUB_OUTPUT"
           echo "ASC version state (pre-metadata): $STATE"
@@ -723,8 +723,8 @@ jobs:
           fi
           python - <<'PY'
           import os
-          from scripts.asc_client import ASCClient
-          from scripts.asc_submit_for_review import attach_build, select_valid_build_id
+          from scripts.asc.asc_client import ASCClient
+          from scripts.asc.asc_submit_for_review import attach_build, select_valid_build_id
 
           client = ASCClient.from_env(timeout=30)
           app_id = os.environ["APP_ID"]
@@ -742,7 +742,7 @@ jobs:
           APPSTORE_ISSUER_ID: ${{ secrets.APPSTORE_ISSUER_ID }}
           APPSTORE_PRIVATE_KEY: /tmp/appstore_key.p8
         run: |
-          python scripts/asc_verify_ready.py \
+          python scripts/asc/asc_verify_ready.py \
             --version "${{ steps.asc_version.outputs.selected_version }}" \
             --json /tmp/asc_ready.json
 
@@ -774,7 +774,7 @@ jobs:
           APPSTORE_PRIVATE_KEY: /tmp/appstore_key.p8
           IOS_VERSION: ${{ steps.asc_version.outputs.selected_version }}
         run: |
-          python scripts/asc_submit_for_review.py \
+          python scripts/asc/asc_submit_for_review.py \
             --version "$IOS_VERSION" \
             --locale en-US \
             --wait
