+++ scripts/asc/asc_add_tester.py
