+++ scripts/tests/test_asc_poll_version_state.py
@@ -11,7 +11,7 @@ class _FakeClient:
 
 class AscPollVersionStateTests(unittest.TestCase):
     def test_find_app_store_version_id_returns_id_and_state(self):
-        from scripts.asc_poll_version_state import find_app_store_version_id
+        from scripts.asc.asc_poll_version_state import find_app_store_version_id
 
         client = _FakeClient(
             [
@@ -28,14 +28,14 @@ class AscPollVersionStateTests(unittest.TestCase):
         self.assertEqual(state, "WAITING_FOR_REVIEW")
 
     def test_find_app_store_version_id_dies_when_missing(self):
-        from scripts.asc_poll_version_state import find_app_store_version_id
+        from scripts.asc.asc_poll_version_state import find_app_store_version_id
 
         client = _FakeClient([])
         with self.assertRaises(SystemExit):
             find_app_store_version_id(client, app_id="app1", version="1.2.3")
 
     def test_find_app_store_version_id_defaults_to_unknown_state(self):
-        from scripts.asc_poll_version_state import find_app_store_version_id
+        from scripts.asc.asc_poll_version_state import find_app_store_version_id
 
         client = _FakeClient([{"id": "ver123", "type": "appStoreVersions"}])
         vid, state = find_app_store_version_id(client, app_id="app1", version="1.2.3")
@@ -43,7 +43,7 @@ class AscPollVersionStateTests(unittest.TestCase):
         self.assertEqual(state, "UNKNOWN")
 
     def test_find_app_store_version_id_dies_when_id_is_missing(self):
-        from scripts.asc_poll_version_state import find_app_store_version_id
+        from scripts.asc.asc_poll_version_state import find_app_store_version_id
 
         client = _FakeClient(
             [
